+++ Tests/SwiftLintFramework/RulesTests.swift
@@ -113,6 +113,10 @@ class RulesTests: XCTestCase {
         verifyRule(EmptyCountRule.description)
     }
 
+    func testExplicitInit() {
+        verifyRule(ExplicitInitRule.description)
+    }
+
     func testFileLength() {
         verifyRule(FileLengthRule.description, commentDoesntViolate: false)
     }